test(seo): keep coverage in source files
Move the article metadata and sitemap regressions into the import.meta.vitest pattern. Include their source modules in Vitest discovery so the checks execute as part of the normal suite.
